+++ workspace/apps/server/src/main.ts
@@ -18,6 +18,7 @@
 import Fastify, { type FastifyInstance } from "fastify";
 
 import { init as initRouters } from "./routers";
+import { init as initPlugins } from "./plugins";
 import { D } from "./models/client";
 import {
     //
@@ -57,6 +58,7 @@ export class Server {
     public async init() {
         await Promise.all([
             initRouters(this.fastify), // 初始化路由
+            initPlugins(this.fastify), // 初始化插件
             D.init(this.fastify), // 初始化数据库
         ]);
         await this.fastify.ready(); // 等待 Fastify 准备就绪