+++ workspace/apps/server/src/plugins/cookie.ts
@@ -15,33 +15,11 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-/**
- * 用户角色
- */
-export enum Role {
-    Administrator = 1, // 管理员
-    Reviewer = 2, // 审核员
-    Staff = 3, // 员工
-    User = 10, // 用户
-    Visitor = 20, // 游客
-}
+import { fastifyCookie } from "@fastify/cookie";
+import type { FastifyInstance } from "fastify";
 
-/**
- * 将角色字符串转换为角色枚举
- * @param role 角色字符串
- */
-export function str2role(role?: string): Role {
-    switch (role) {
-        case "administrator":
-            return Role.Administrator;
-        case "reviewer":
-            return Role.Reviewer;
-        case "staff":
-            return Role.Staff;
-        case "user":
-            return Role.User;
-        case "visitor":
-        default:
-            return Role.Visitor;
-    }
+export async function register(fastify: FastifyInstance) {
+    // REF: https://www.npmjs.com/package/@fastify/jwt?activeTab=readme#example-using-cookie
+    await fastify.register(fastifyCookie);
+    await fastify.after();
 }